train.py
"""
训练脚本
"""
import argparse
import yaml
import torch
from torch.utils.data import DataLoader
from torchvision import transforms
from tqdm import tqdm
import os
from pathlib import Path

from src.data.dataset import NYUDepthV2Dataset
from src.models.blip2 import BLIP2Model
from src.models.grounding_dino import GroundingDINOModel


def train_grounding(model, dataloader, optimizer, device, epoch):
    """训练 Grounding 任务"""
    model.model.train()
    total_loss = 0
    
    for batch in tqdm(dataloader, desc=f"Epoch {epoch} - Grounding"):
        images = batch['image'].to(device)
        # 这里需要根据实际任务设计损失函数
        # 简化版本，实际需要根据 Grounding DINO 的训练方式调整
        pass
    
    return total_loss / len(dataloader)


def train_vqa(model, dataloader, optimizer, device, epoch):
    """训练 VQA 任务"""
    model.model.train()
    total_loss = 0
    
    for batch in tqdm(dataloader, desc=f"Epoch {epoch} - VQA"):
        images = batch['image'].to(device)
        # 这里需要根据实际任务设计损失函数
        pass
    
    return total_loss / len(dataloader)


def main():
    parser = argparse.ArgumentParser(description="训练家居机器人模型")
    parser.add_argument("--config", type=str, default="config.yaml",
                       help="配置文件路径")
    args = parser.parse_args()
    
    # 加载配置
    with open(args.config, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    
    device = config['model']['device']
    if device == "cuda" and not torch.cuda.is_available():
        print("CUDA 不可用，使用 CPU")
        device = "cpu"
    
    # 创建数据加载器
    transform = transforms.Compose([
        transforms.Resize(config['data']['image_size']),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406],
                           std=[0.229, 0.224, 0.225])
    ])
    
    train_dataset = NYUDepthV2Dataset(
        data_path=config['data']['dataset_path'],
        split="train",
        transform=transform,
        image_size=config['data']['image_size']
    )
    
    train_loader = DataLoader(
        train_dataset,
        batch_size=config['data']['batch_size'],
        shuffle=True,
        num_workers=config['data']['num_workers']
    )
    
    # 初始化模型
    if config['tasks']['vqa']['enabled']:
        vqa_model = BLIP2Model(
            model_name=config['model']['blip2_model'],
            device=device,
            precision=config['model']['precision']
        )
    
    if config['tasks']['grounding']['enabled']:
        grounding_model = GroundingDINOModel(
            model_path=config['model']['grounding_model'],
            device=device
        )
    
    # 创建保存目录
    save_dir = Path(config['training']['save_dir'])
    save_dir.mkdir(parents=True, exist_ok=True)
    
    print("=" * 50)
    print("开始训练")
    print("=" * 50)
    print(f"设备: {device}")
    print(f"训练样本数: {len(train_dataset)}")
    print(f"批次大小: {config['data']['batch_size']}")
    print(f"训练轮数: {config['training']['num_epochs']}")
    print("=" * 50)
    
    # 训练循环
    for epoch in range(1, config['training']['num_epochs'] + 1):
        print(f"\nEpoch {epoch}/{config['training']['num_epochs']}")
        
        # 这里需要根据实际需求实现训练逻辑
        # 由于 Grounding DINO 和 BLIP-2 通常是预训练模型，
        # 这里主要是微调或端到端训练
        
        if epoch % config['training']['save_every'] == 0:
            # 保存检查点
            checkpoint = {
                'epoch': epoch,
                'config': config
            }
            torch.save(checkpoint, save_dir / f"checkpoint_epoch_{epoch}.pth")
            print(f"检查点已保存: {save_dir / f'checkpoint_epoch_{epoch}.pth'}")
    
    print("\n训练完成!")


if __name__ == "__main__":
    main()

