requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
timm>=0.9.0
mmcv>=2.0.0
mmdet>=3.0.0
mmengine>=0.8.0
groundingdino-py>=0.4.0
sentencepiece>=0.1.99
accelerate>=0.20.0
bitsandbytes>=0.39.0
peft>=0.4.0
einops>=0.6.0
decord>=0.6.0
scipy>=1.10.0
h5py>=3.8.0
scikit-learn>=1.3.0
jupyter>=1.0.0
ipywidgets>=8.0.0


